CHANGE-1:
Separated problems into different folders where each folder is having three files:
    fileName.cpp
    input.txt
    output.txt

Example:
    Contest Name: checkCodemon
    then the structure of the folder is as follows:
        checkCodemon --A-|-A.cpp
                         |-input.txt
                         |-output.txt
                     --B-|-B.cpp
                         |-input.txt
                         |-output.txt ................

CodemonListen.py is also edited so as to recognize the path of .cpp file during compilation and running.

CHANGE-2:
If the arguements are 'codemon init', then there is an error of 'index out of range'.
To overcome this problem, an if condition is added in codemon.py.